fixes for windows

+++ core/src/main/java/com/crawljax/browser/WebDriverBackedEmbeddedBrowser.java
@@ -943,13 +943,10 @@ public final class WebDriverBackedEmbeddedBrowser implements EmbeddedBrowser {
 
         if (os.contains("win")) {
 
-            Long width = (Long) chromeDriver.executeScript("return document.body.scrollWidth");
-            Long height = (Long) chromeDriver.executeScript("return document.body.scrollHeight");
-
-            // Correctly handle devicePixelRatio as Double
-            Double scale = (Double) chromeDriver.executeScript("return window.devicePixelRatio");
-
-            int intscale = scale.intValue();
+            long width = (long) chromeDriver.executeScript("return document.body.scrollWidth");
+            long height = (long) chromeDriver.executeScript("return document.body.scrollHeight");
+            long scale = (long) chromeDriver.executeScript("return window.devicePixelRatio");
+            int intscale = (int) scale;
 
             HashMap<String, Object> setDeviceMetricsOverride = new HashMap<>();
             setDeviceMetricsOverride.put("deviceScaleFactor", scale);
